+++ src/main/java/dinkplugin/notifiers/SlayerNotifier.java
@@ -9,6 +9,7 @@ import org.jetbrains.annotations.VisibleForTesting;
 
 import javax.inject.Singleton;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -19,7 +20,7 @@ public class SlayerNotifier extends BaseNotifier {
     static final Pattern SLAYER_TASK_REGEX = Pattern.compile("You have completed your task! You killed (?<task>[\\d,]+ [^.]+)\\..*");
     private static final Pattern SLAYER_COMPLETE_REGEX = Pattern.compile("You've completed (?:at least )?(?<taskCount>[\\d,]+) (?:Wilderness )?tasks?(?: and received (?<points>[\\d,]+) points, giving you a total of [\\d,]+|\\.You'll be eligible to earn reward points if you complete tasks from a more advanced Slayer Master\\.| and reached the maximum amount of Slayer points \\((?<points2>[\\d,]+)\\))?");
 
-    private volatile String slayerTask = "";
+    private final AtomicReference<String> slayerTask = new AtomicReference<>("");
     private final AtomicInteger badTicks = new AtomicInteger(); // used to prevent notifs from using stale data
 
     @Override
@@ -34,11 +35,11 @@ public class SlayerNotifier extends BaseNotifier {
 
     public void onChatMessage(String chatMessage) {
         if (isEnabled()) {
-            if (slayerTask.isEmpty()) {
+            if (slayerTask.get().isEmpty()) {
                 Matcher bossMatcher = BOSS_REGEX.matcher(chatMessage);
                 if (bossMatcher.find()) {
                     String name = bossMatcher.group("name");
-                    this.slayerTask = name.endsWith(" boss") ? name.substring(0, name.length() - " boss".length()) : name;
+                    this.slayerTask.set(name.endsWith(" boss") ? name.substring(0, name.length() - " boss".length()) : name);
                     return;
                 }
             }
@@ -46,15 +47,15 @@ public class SlayerNotifier extends BaseNotifier {
             Matcher taskMatcher = SLAYER_TASK_REGEX.matcher(chatMessage);
             if (taskMatcher.find()) {
                 String task = taskMatcher.group("task");
-                if (slayerTask.isEmpty()) {
-                    this.slayerTask = task;
-                } else {
-                    this.slayerTask = String.format("%s %s", task.substring(0, task.indexOf(' ')), slayerTask);
-                }
+                slayerTask.getAndUpdate(old -> {
+                    if (old == null || old.isEmpty())
+                        return task;
+                    return String.format("%s %s", task.substring(0, task.indexOf(' ')), old);
+                });
                 return;
             }
 
-            if (slayerTask.isEmpty()) {
+            if (slayerTask.get().isEmpty()) {
                 return;
             }
 
@@ -79,7 +80,7 @@ public class SlayerNotifier extends BaseNotifier {
 
     public void onTick() {
         // Track how many ticks occur where we only have partial slayer data
-        if (!slayerTask.isEmpty())
+        if (!slayerTask.get().isEmpty())
             badTicks.getAndIncrement();
 
         // Clear data if 2 ticks pass with only partial parsing
@@ -88,7 +89,8 @@ public class SlayerNotifier extends BaseNotifier {
     }
 
     private void handleNotify(String slayerPoints, String slayerCompleted) {
-        if (slayerPoints.isEmpty() || slayerTask.isEmpty() || slayerCompleted.isEmpty()) {
+        String task = slayerTask.get();
+        if (task.isEmpty() || slayerPoints.isEmpty() || slayerCompleted.isEmpty()) {
             return;
         }
 
@@ -97,12 +99,12 @@ public class SlayerNotifier extends BaseNotifier {
             String notifyMessage = StringUtils.replaceEach(
                 config.slayerNotifyMessage(),
                 new String[] { "%USERNAME%", "%TASK%", "%TASKCOUNT%", "%POINTS%" },
-                new String[] { Utils.getPlayerName(client), slayerTask, slayerCompleted, slayerPoints }
+                new String[] { Utils.getPlayerName(client), task, slayerCompleted, slayerPoints }
             );
 
             createMessage(config.slayerSendImage(), NotificationBody.builder()
                 .text(notifyMessage)
-                .extra(new SlayerNotificationData(slayerTask, slayerCompleted, slayerPoints))
+                .extra(new SlayerNotificationData(task, slayerCompleted, slayerPoints))
                 .type(NotificationType.SLAYER)
                 .build());
         }
@@ -111,7 +113,7 @@ public class SlayerNotifier extends BaseNotifier {
     }
 
     public void reset() {
-        slayerTask = "";
+        slayerTask.set("");
         badTicks.set(0);
     }
 }