+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Minor: Include number of completed entries in collection log notification. (#174, #181)
 - Bugfix: Include PKer from Last Man Standing minigame in death notifications. (#187)
 - Bugfix: Fire slayer notifications beyond 999 points. (#182)
+- Dev: Improve thread-safety of slayer notifier. (#184)
 - Dev: Update Gradle to 8.0.2 patch version. (#185)
 - Dev: Add issue tracker link to README. (#179)